refactor: Mock Data 분리 및 디렉토리 이동
