+++ client/src/libs/cache/aggregateCategoryCache/aggregateCategoryData.test.ts
@@ -1,38 +1,6 @@
 import { fail } from 'assert';
+import { mockAggregateCategoryData } from '@/commons/mockData';
 import aggregateCategoryDataCache from '.';
-
-const mockAggregateCategoryData = {
-  income: [
-    {
-      category: '월급',
-      aggregate: '300000',
-      dataArray: [
-        {
-          tid: 1,
-          amount: 300000,
-          tradeAt: new Date(),
-          description: 'mock transaction',
-          payment: '신한은행',
-        },
-      ],
-    },
-  ],
-  expenditure: [
-    {
-      category: '식비',
-      aggregate: '3000',
-      dataArray: [
-        {
-          tid: 1,
-          amount: 3000,
-          tradeAt: new Date(),
-          description: 'mock transaction',
-          payment: '신한체크',
-        },
-      ],
-    },
-  ],
-};
 
 it('응답데이터 캐싱 테스트', () => {
   const date = { year: 2020, month: 12 };
