+++ client/src/libs/cache/categoryCache/categoryCache.test.ts
@@ -1,21 +1,17 @@
+import { mockCategories } from '@/commons/mockData';
 import categoryCache from '.';
-
-const mockResponse = [
-  { cid: 1, name: '식비', isIncome: false, uid: 1 },
-  { cid: 2, name: '교통', isIncome: false, uid: 1 },
-];
 
 afterEach(() => {
   categoryCache.clear();
 });
 
 it('응답데이터 캐싱 테스트', () => {
-  categoryCache.set(mockResponse);
+  categoryCache.set(mockCategories);
   expect(categoryCache.get().length).toBe(2);
 });
 
 it('캐시 clear 테스트', () => {
-  categoryCache.set(mockResponse);
+  categoryCache.set(mockCategories);
   expect(categoryCache.get().length).toBe(2);
 
   categoryCache.clear();
@@ -23,7 +19,7 @@ it('캐시 clear 테스트', () => {
 });
 
 it('반환 데이터 immutable 테스트', () => {
-  categoryCache.set(mockResponse);
+  categoryCache.set(mockCategories);
   const cachedData = categoryCache.get();
 
   categoryCache.clear();
